split graph tests into two files
