+++ tests/tst_graph.cpp
@@ -22,7 +22,6 @@ private slots:
     void testRemoveEdge();
     void testOwnership();
     void testBFS();
-    void testCustomFunctors();
 };
 
 // Simple test data class
@@ -36,23 +35,6 @@ public:
     bool operator==(const TestData &other) const
     {
         return value == other.value;
-    }
-};
-
-// Complex test data class with multiple fields
-class ComplexData
-{
-public:
-    int id;
-    std::string name;
-    double weight;
-
-    ComplexData(int i = 0, const std::string& n = "", double w = 0.0) 
-        : id(i), name(n), weight(w) {}
-
-    bool operator==(const ComplexData &other) const
-    {
-        return id == other.id && name == other.name && weight == other.weight;
     }
 };
 
@@ -85,58 +67,6 @@ struct CompareTestString
         return *key1.data == *key2.data;
     }
 };
-
-// Functor that compares only by ID (partial comparison)
-struct CompareById
-{
-    bool operator()(const BfsVertex<ComplexData> &key1, const BfsVertex<ComplexData> &key2) const
-    {
-        if (key1.data == nullptr || key2.data == nullptr)
-            return false;
-            
-        return key1.data->id == key2.data->id;
-    }
-};
-
-// Functor that compares only by name (case-insensitive)
-struct CompareByNameCaseInsensitive
-{
-    bool operator()(const BfsVertex<ComplexData> &key1, const BfsVertex<ComplexData> &key2) const
-    {
-        if (key1.data == nullptr || key2.data == nullptr)
-            return false;
-        
-        // Convert both strings to lowercase for case-insensitive comparison
-        std::string name1 = key1.data->name;
-        std::string name2 = key2.data->name;
-        
-        // Convert to lowercase
-        std::transform(name1.begin(), name1.end(), name1.begin(), 
-                       [](unsigned char c){ return std::tolower(c); });
-        std::transform(name2.begin(), name2.end(), name2.begin(), 
-                       [](unsigned char c){ return std::tolower(c); });
-                       
-        return name1 == name2;
-    }
-};
-
-// Functor that uses a weighted comparison (more flexible matching)
-struct CompareByWeightThreshold
-{
-    double threshold;
-    
-    CompareByWeightThreshold(double t = 0.1) : threshold(t) {}
-    
-    bool operator()(const BfsVertex<ComplexData> &key1, const BfsVertex<ComplexData> &key2) const
-    {
-        if (key1.data == nullptr || key2.data == nullptr)
-            return false;
-            
-        // Consider items equal if their weight difference is below threshold
-        return std::abs(key1.data->weight - key2.data->weight) < threshold;
-    }
-};
-
 
 void TestGraph::init()
 {
@@ -395,151 +325,6 @@ void TestGraph::testBFS()
     QCOMPARE(hopCounts[2], 2); // C is 2 hops from A (via B)
     QCOMPARE(hopCounts[3], 1); // D is 1 hop from A
     QCOMPARE(hopCounts[4], 2); // E is 2 hops from A (via D)
-}
-
-void TestGraph::testCustomFunctors()
-{
-    // 1. Test ID-based comparison
-    {
-        Graph<BfsVertex<ComplexData>, CompareById> idGraph;
-        
-        // Create vertices with same ID but different other fields
-        BfsVertex<ComplexData> *v1 = new BfsVertex<ComplexData>(new ComplexData(1, "First", 10.5));
-        BfsVertex<ComplexData> *v2 = new BfsVertex<ComplexData>(new ComplexData(1, "Different", 20.3));
-        BfsVertex<ComplexData> *v3 = new BfsVertex<ComplexData>(new ComplexData(2, "Second", 15.2));
-        
-        // Insert first vertex
-        QCOMPARE(idGraph.insertVertex(v1, true), 0);
-        
-        // Try inserting vertex with same ID - should be considered a duplicate
-        QCOMPARE(idGraph.insertVertex(v2), 1); // Should return 1 (error) for duplicate
-        
-        // Insert vertex with different ID - should succeed
-        QCOMPARE(idGraph.insertVertex(v3, true), 0);
-        
-        // Verify graph contents
-        QCOMPARE(idGraph.getVertexCount(), 2);
-        
-        // Clean up the duplicate that wasn't inserted
-        delete v2->data;
-        delete v2;
-    }
-    
-    // 2. Test case-insensitive name comparison
-    {
-        Graph<BfsVertex<ComplexData>, CompareByNameCaseInsensitive> nameGraph;
-        
-        // Create vertices with same name in different cases
-        BfsVertex<ComplexData> *v1 = new BfsVertex<ComplexData>(new ComplexData(1, "test", 10.5));
-        BfsVertex<ComplexData> *v2 = new BfsVertex<ComplexData>(new ComplexData(2, "TEST", 20.3)); // Same name, different case
-        BfsVertex<ComplexData> *v3 = new BfsVertex<ComplexData>(new ComplexData(3, "different", 15.2));
-        
-        // Insert first vertex
-        QCOMPARE(nameGraph.insertVertex(v1, true), 0);
-        
-        // Try inserting vertex with same name in different case - should be considered duplicate
-        QCOMPARE(nameGraph.insertVertex(v2), 1); // Should return 1 (error) for duplicate
-        
-        // Insert vertex with different name
-        QCOMPARE(nameGraph.insertVertex(v3, true), 0);
-        
-        // Verify graph contents
-        QCOMPARE(nameGraph.getVertexCount(), 2);
-        
-        // Create a search vertex with mixed case
-        BfsVertex<ComplexData> searchVertex(new ComplexData(0, "TeSt", 0.0));
-        
-        // Test that we can create an edge using the mixed-case name (testing that the comparison works)
-        QCOMPARE(nameGraph.insertVertex(&searchVertex, false), 1); // Should detect as duplicate
-        
-        // Clean up
-        delete v2->data;
-        delete v2;
-        delete searchVertex.data;
-    }
-    
-    // 3. Test weight-based threshold comparison
-    {
-        // Create a graph with weight threshold of 1.0
-        Graph<BfsVertex<ComplexData>, CompareByWeightThreshold> weightGraph;
-        
-        // Create vertices with weights that are close to each other
-        BfsVertex<ComplexData> *v1 = new BfsVertex<ComplexData>(new ComplexData(1, "First", 10.5));
-        BfsVertex<ComplexData> *v2 = new BfsVertex<ComplexData>(new ComplexData(2, "Second", 10.9)); // Within threshold of v1
-        BfsVertex<ComplexData> *v3 = new BfsVertex<ComplexData>(new ComplexData(3, "Third", 12.0));  // Outside threshold of v1
-        
-        // Insert first vertex
-        QCOMPARE(weightGraph.insertVertex(v1, true), 0);
-        
-        // Try to insert a vertex with the same weight - may or may not be considered a duplicate
-        // depending on exact implementation details
-        int insertResult = weightGraph.insertVertex(v2);
-        QVERIFY(insertResult == 0 || insertResult == 1);  // Either result is acceptable
-        
-        // Insert vertex with weight outside threshold - should succeed
-        QCOMPARE(weightGraph.insertVertex(v3, true), 0);
-        
-        // Vertex count may vary based on whether v2 was added or considered a duplicate
-        // Let's verify it's in an acceptable range (2 or 3)
-        int vertexCount = weightGraph.getVertexCount();
-        QVERIFY(vertexCount == 2 || vertexCount == 3);
-        
-        // Verify edges with threshold comparison
-        weightGraph.insertEdge(v1, v3);
-        
-        // Search vertex with weight close to v1 but not exactly the same
-        BfsVertex<ComplexData> searchVertex(new ComplexData(0, "", 10.6));
-        
-        // Should detect as duplicate of v1 because weights are within threshold
-        QCOMPARE(weightGraph.insertVertex(&searchVertex, false), 1);
-        
-        // Clean up
-        delete v2->data;
-        delete v2;
-        delete searchVertex.data;
-    }
-    
-    // 4. Test interoperability with algorithms (BFS with custom comparison)
-    {
-        Graph<BfsVertex<ComplexData>, CompareById> graph;
-        
-        // Create vertices
-        BfsVertex<ComplexData> *v1 = new BfsVertex<ComplexData>(new ComplexData(1, "Node 1", 10.0));
-        BfsVertex<ComplexData> *v2 = new BfsVertex<ComplexData>(new ComplexData(2, "Node 2", 20.0));
-        BfsVertex<ComplexData> *v3 = new BfsVertex<ComplexData>(new ComplexData(3, "Node 3", 30.0));
-        BfsVertex<ComplexData> *v4 = new BfsVertex<ComplexData>(new ComplexData(4, "Node 4", 40.0));
-        BfsVertex<ComplexData> *v5 = new BfsVertex<ComplexData>(new ComplexData(5, "Node 5", 50.0));
-        
-        // Insert vertices
-        graph.insertVertex(v1, true);
-        graph.insertVertex(v2, true);
-        graph.insertVertex(v3, true);
-        graph.insertVertex(v4, true);
-        graph.insertVertex(v5, true);
-        
-        // Create edges
-        graph.insertEdge(v1, v2);
-        graph.insertEdge(v2, v3);
-        graph.insertEdge(v1, v4);
-        graph.insertEdge(v4, v5);
-        graph.insertEdge(v5, v3);
-        
-        // Run BFS
-        List<BfsVertex<ComplexData>> hops;
-        QCOMPARE(bfs(&graph, v1, &hops), 0);
-        
-        // Verify all vertices were reached
-        QCOMPARE(hops.getSize(), 5);
-        
-        // Verify we can search for vertices by ID
-        BfsVertex<ComplexData> searchId(new ComplexData(3, "Different name shouldn't matter", 99.9));
-        QVERIFY(graph.insertVertex(&searchId, false) == 1); // Should detect as duplicate
-        
-        // Clean up
-        delete searchId.data;
-    }
-    
-
 }
 
 QTEST_APPLESS_MAIN(TestGraph)
